Tp03 juri enzo (#120)
* TP02 - Relaciones y DTO implementados

* TP03: HTTP y errores

* TP03: HTTP y errores sin try-catch

---------

+++ src/main/java/com/imb2025/smedico/service/jpa/AfiliacionServiceImpl.java
@@ -22,19 +22,19 @@ public class AfiliacionServiceImpl implements IAfiliacionService {
 	private ObraSocialRepository repoobra;
 
 	@Override
-	public List<Afiliacion> findAll() {
+	public List<Afiliacion> findAll() throws Exception {		
 		return repo.findAll();
 	}
 
 	@Override
-	public Afiliacion findById(Long id) {
+	public Afiliacion findById(Long id) throws Exception {
 		Optional<Afiliacion> opt;
 		opt = repo.findById(id);
 
 		if (opt.isPresent()) {
 			return opt.get();
 		} else {
-			return null;
+			throw new Exception("Afiliación no encontrada");
 		}
 	}
 
@@ -54,24 +54,22 @@ public class AfiliacionServiceImpl implements IAfiliacionService {
 	}
 
 	@Override
-	public void deleteById(Long id) {
-		repo.deleteById(id);
+	public void deleteById(Long id) throws Exception {
+		if (repo.existsById(id)) {
+
+			repo.deleteById(id);
+		} else {
+			throw new Exception("Afiliación no encontrada");
+		}
+
 	}
 
-	public Afiliacion  fromDto(AfiliacionRequestDTO dto) {
-		 try {
-			 Afiliacion fili = new Afiliacion(
-					dto.getNumeroAfiliado(),
-					dto.getFechaVigenciaDesde(),
-					dto.getFechaHasta(),		            
-					repopaciente.findById(dto.getIdpaciente()).orElseThrow(() -> new Exception("Paciente no encontrado")),
-					repoobra.findById(dto.getIdobra()).orElseThrow(() -> new Exception("Obra social no encontrada"))					
-		        );
+	public Afiliacion fromDto(AfiliacionRequestDTO dto) throws Exception {
+		Afiliacion fili = new Afiliacion(dto.getNumeroAfiliado(), dto.getFechaVigenciaDesde(), dto.getFechaHasta(),
+				repopaciente.findById(dto.getIdpaciente()).orElseThrow(() -> new Exception("Paciente no encontrado")),
+				repoobra.findById(dto.getIdobra()).orElseThrow(() -> new Exception("Obra social no encontrada")));
 
-			 return fili;
-		    } catch (Exception e) {
-		    	System.err.println("Error al convertir el DTO a Afiliacion: " + e.getMessage());
-		    	throw new RuntimeException("Hubo un error: " + e.getMessage(), e);
-				}
-			}
+		return fili;
+
+	}
 }
